Unify the View Row/Column implementations
They're 99% the same. That also means we can fold in the helper.
